% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-chr_to_num.R
\name{chr_to_num}
\alias{chr_to_num}
\title{Clean and Possibly Convert Numbers in \code{character} Representation}
\usage{
chr_to_num(
  x,
  std = TRUE,
  warn = TRUE,
  convert = TRUE,
  na = na_patterns,
  replace = data.frame(pattern = c("MRD BY NGS ?", "^NEGATIVE$",
    "^(?:WE[EA]KLY )?POSITIVE$", ".*\\\\bIN CR\\\\b.*"), replacement = c("", "0", ">0",
    "0")),
  per_action = c("drop", "divide", "ignore"),
  multiple_decimals = c("use_first", "use_last", "ignore"),
  donor_host = c("use_donor", "use_host", "ignore")
)
}
\arguments{
\item{x}{A \code{character} vector}

\item{std}{Whether to standardize the vector before cleaning and converting}

\item{convert}{Whether to actually convert to \code{numeric}}

\item{na}{Regular expressions to convert to \code{NA}}

\item{replace}{A \code{data.frame} of regular expressions and strings to replace
them; regular expression should be in a column named \code{pattern}, and
replacements should be in a column named \code{replacement}. Each row is passed
to \code{stringr::str_replace()}.}

\item{per_action}{How to treat \%/percent/per million/etc labels. \code{drop} simply
removes the labels, \code{divide} divides the value by the appropriate denominator,
and \code{ignore} does nothing.}

\item{multiple_decimals}{How to handle multiple decimals within a number}

\item{donor_host}{Which value to use when values for both a donor and a host
are given}
}
\value{
A \code{numeric} or \code{character} vector, depending on the value of \code{convert}
}
\description{
\code{chr_to_num()} cleans strings representing numeric values. It is tailored for
use with the ML in HCT dataset; use with other data may not go as expected.
}
\details{
The function first converts strings matching \code{na_patterns} to missing values.
It then simplifies any numeric representations it finds, including 10^x and
various common typos observed in the data (unneeded decimals, commas, and zeros).
Additionally, it converts Excel datetimes of the form 1/x/1900 hh:mm back to
decimal representation, and additionally converts fractions less than 1 to
decimals. Lastly, it handles a few idiosyncratic text strings, including
conversion of POSITIVE and NEGATIVE values, as well as W(E|A)KLY POSITIVE
and IN CR (complete remission). Optionally, it will extract values labelled as
donor or host (patient), which is specific to the chimerism dataset. It also
removes the leading text MRD BY NGS. By default, the function emits a warning
when potential numeric values are not able to be converted to \code{numeric}.
}
\keyword{internal}
