% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_server.R
\name{con_sql_server}
\alias{con_sql_server}
\title{Connect to SQL Server Where Data is Stored}
\usage{
con_sql_server(dbname = c("IRB_MLinHCT", "EDW"))
}
\arguments{
\item{dbname}{\verb{[chr(1)]} The name of the database to connect to}
}
\value{
\verb{[Microsoft SQL Server]} An ODBC connection object
}
\description{
Connect to SQL Server Where Data is Stored
}
