% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_master.R
\name{convert_degree_match}
\alias{convert_degree_match}
\title{Convert Degree of Match \code{character} Data to Count with Specified Denominator}
\usage{
convert_degree_match(x, denom = 6L)
}
\arguments{
\item{x}{\verb{[character]} Vector of degree of match counts}

\item{denom}{\verb{[integer(1)]} Denominator of counts to extract}
}
\value{
An \code{integer} vector of counts
}
\description{
Convert Degree of Match \code{character} Data to Count with Specified Denominator
}
\keyword{internal}
