% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{df_class}
\alias{df_class}
\title{Get Class of \code{data.frame}-Like Object}
\usage{
df_class(data)
}
\arguments{
\item{data}{\verb{[data.frame]} A \code{data.frame} or one of its sub-classes}
}
\value{
\verb{[chr(1)]} The name of the class
}
\description{
Get Class of \code{data.frame}-Like Object
}
\keyword{internal}
