% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_cache.R
\name{dm_cache_check}
\alias{dm_cache_check}
\title{Check Hash for \code{dm} Object}
\usage{
dm_cache_check(
  checksum,
  file,
  dir = path_create("~", ".cache_dmhct"),
  n_threads = max(1L, parallel::detectCores()\%/\%2L)
)
}
\arguments{
\item{checksum}{The hash of the \code{dm} object}

\item{file}{The file stem to use for the cache objects}

\item{dir}{The directory in which to save the cache objects}

\item{n_threads}{The number of threads to use when serializing the \code{dm}}
}
\value{
A \code{logical} indicating whether the passed \code{checksum} and cached
\code{checksum} are identical
}
\description{
Check Hash for \code{dm} Object
}
\keyword{internal}
