% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_cache.R
\name{dm_cache_read}
\alias{dm_cache_read}
\title{Read \code{dm} Data or its Checksum from a the Specified Cache}
\usage{
dm_cache_read(
  obj = c("data", "checksum"),
  file,
  dir = path_create("~", ".cache_dmhct"),
  fail = TRUE,
  n_threads = if (obj == "data") max(1L, parallel::detectCores()\%/\%2L) else 1L
)
}
\arguments{
\item{obj}{What to read from the cache}

\item{file}{The file stem to use for the cache objects}

\item{dir}{The directory in which to save the cache objects}

\item{fail}{Whether to error if \code{dir}/\code{obj}_\code{file} combo does not exist}

\item{n_threads}{The number of threads to use when serializing the \code{dm}}
}
\value{
The de-serialized object
}
\description{
Read \code{dm} Data or its Checksum from a the Specified Cache
}
\keyword{internal}
