% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_cache.R
\name{dm_cache_write}
\alias{dm_cache_write}
\title{Cache a \code{dm} Object in the Specified File and Directory}
\usage{
dm_cache_write(
  dm,
  checksum,
  file,
  dir = path_create(system.file("", package = "dmhct"), "cache"),
  n_threads = max(1L, parallel::detectCores()\%/\%2L)
)
}
\arguments{
\item{dm}{A \code{dm} object}

\item{checksum}{The hash of the \code{dm} object}

\item{file}{The file stem to use for the cache objects}

\item{dir}{The directory in which to save the cache objects}

\item{n_threads}{The number of threads to use when serializing the \code{dm}}
}
\value{
The \code{dm} (invisibly)
}
\description{
Cache a \code{dm} Object in the Specified File and Directory
}
\keyword{internal}
