% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_cerner.R
\name{dm_cerner_extract}
\alias{dm_cerner_extract}
\title{Unite All Cerner Tables and Remove Individual Tables}
\usage{
dm_cerner_extract(
  dm_remote,
  path_var = system.file("extdata", "cerner_var_map.csv", package = "predHCT"),
  quiet = TRUE
)
}
\arguments{
\item{dm_remote}{\verb{[dm]} Remote \code{dm} connected to an SQL Server w/ HLA data}

\item{path_var}{\verb{[chr(1)]} Path to CSV file mapping test names to
standardized variable names}

\item{quiet}{Should the function return quietly if a Cerner table already
exists?}
}
\value{
The updated \code{dm}
}
\description{
Unite All Cerner Tables and Remove Individual Tables
}
