% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_cerner.R
\name{dm_cerner_transform}
\alias{dm_cerner_transform}
\title{Transform an Extracted Cerner Table in a Local \code{dm}}
\usage{
dm_cerner_transform(
  dm_local,
  numeric_only = TRUE,
  path = system.file("extdata", "cerner_var_map.csv", package = "dmhct")
)
}
\arguments{
\item{dm_local}{\verb{[dm]} Local \code{dm} w/ HCT data}

\item{numeric_only}{\verb{[lgl(1)]} Should non-numeric data be dropped during
transformation?}

\item{path}{\verb{[chr(1)]} Path to CSV file mapping test names to
standardized variable names}
}
\value{
The \code{dm} object w/ transformed \code{cerner} table
}
\description{
Transform an Extracted Cerner Table in a Local \code{dm}
}
