% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_cache.R
\name{dm_checksum}
\alias{dm_checksum}
\title{Hash a \code{dm} Object}
\usage{
dm_checksum(dm)
}
\arguments{
\item{dm}{The \code{dm} object to hash}
}
\value{
The hash string of the object
}
\description{
\code{dm_checksum()} creates a checksum by hashing each table + its name individually,
sorting by name, and hashing a list of the results. This means that the hash
does not change when the ordering of the tables change in the dm, but does
change if any data or names are added, deleted, or modified.
}
\keyword{internal}
