% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_chimerism.R
\name{dm_chimerism_extract}
\alias{dm_chimerism_extract}
\title{Extract Chimerism Table from SQL Server}
\usage{
dm_chimerism_extract(dm_remote)
}
\arguments{
\item{dm_remote}{\verb{[dm]} Remote \code{dm} connected to SQL Server w/ HCT data}
}
\value{
\verb{[dm]} The \code{dm} w/ instructions for updating the \code{chimerism} table
}
\description{
Extract Chimerism Table from SQL Server
}
