% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-dm.R
\name{dm_collect}
\alias{dm_collect}
\title{Collect All Tables in a \code{dm} Object}
\usage{
dm_collect(dm_remote, data_table = FALSE)
}
\arguments{
\item{dm_remote}{\verb{[dm]} A \code{dm} object connected to a remote source}

\item{data_table}{\verb{[lgl]} Whether to return a \code{data.table}. If \code{FALSE},
(the default), will return a \code{tibble} instead.}
}
\value{
\verb{[dm]} A new \code{dm} object containing the collected (local) tables
}
\description{
Collect All Tables in a \code{dm} Object
}
