% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_combine.R
\name{dm_combine}
\alias{dm_combine}
\title{Combine Table with Like Information}
\usage{
dm_combine(dm_std = dm_standardize(), quiet = FALSE)
}
\arguments{
\item{dm_std}{A standardized \code{dm} object. Standardization is necessary to
ensure columns are all of the same type.}

\item{quiet}{Should update messages be suppressed?}
}
\value{
The updated \code{dm} object
}
\description{
Combine Table with Like Information
}
