% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_death.R
\name{dm_death_extract}
\alias{dm_death_extract}
\title{Extract Death Table from SQL Server}
\usage{
dm_death_extract(dm_remote)
}
\arguments{
\item{dm_remote}{\verb{[dm]} Remote \code{dm} connected to SQL Server w/ HCT data}
}
\value{
\verb{[dm]} The \code{dm} with instructions for updating the \code{death} table
}
\description{
Extract Death Table from SQL Server
}
