% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_death.R
\name{dm_death_transform}
\alias{dm_death_transform}
\title{Transform the Death Table in a Local \code{dm}}
\usage{
dm_death_transform(dm_local)
}
\arguments{
\item{dm_local}{\verb{[dm]} Local \code{dm} with HCT data}
}
\value{
\verb{[dm]} The \code{dm} with updated \code{death} table
}
\description{
Transform the Death Table in a Local \code{dm}
}
