% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dm_disconnect}
\alias{dm_disconnect}
\title{Disconnect a \code{dm} Object from the Remote Server}
\usage{
dm_disconnect(dm)
}
\arguments{
\item{dm}{The \code{dm} object to disconnect}
}
\value{
The \code{dm} (invisibly)
}
\description{
Disconnect a \code{dm} Object from the Remote Server
}
\keyword{internal}
