% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_disease_status.R
\name{dm_disease_status_extract}
\alias{dm_disease_status_extract}
\title{Extract Disease Status Table from SQL Server}
\usage{
dm_disease_status_extract(dm_remote)
}
\arguments{
\item{dm_remote}{\verb{[dm]} Remote \code{dm} object connected to SQL Server w/ HCT data}
}
\value{
\verb{[dm]} The \code{dm} object w/ instructions for updating \code{disease_status} table
}
\description{
Extract Disease Status Table from SQL Server
}
