% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_disease_status.R
\name{dm_disease_status_transform}
\alias{dm_disease_status_transform}
\title{Transform Disease Status Table in Local \code{dm}}
\usage{
dm_disease_status_transform(dm_local)
}
\arguments{
\item{dm_local}{\code{dm} object}
}
\value{
The \code{dm} object with transformed \code{disease_status} table
}
\description{
Transform Disease Status Table in Local \code{dm}
}
