% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_engraftment.R
\name{dm_engraftment_extract}
\alias{dm_engraftment_extract}
\title{Extract Engraftment Table from SQL Server}
\usage{
dm_engraftment_extract(dm_remote)
}
\arguments{
\item{dm_remote}{\verb{[dm]} A remote \code{dm} connected to an SQL Server w/ HLA data}
}
\value{
The \code{dm} with instructions for updating the \code{engraftment} table
}
\description{
Extract Engraftment Table from SQL Server
}
