% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_engraftment.R
\name{dm_engraftment_transform}
\alias{dm_engraftment_transform}
\title{Transform the Engraftment Table in a Local \code{dm}}
\usage{
dm_engraftment_transform(dm_local)
}
\arguments{
\item{dm_local}{\verb{[dm]} A \code{dm} with HLA data}
}
\value{
The \code{dm} with the \code{engraftment} table transformed
}
\description{
Transform the Engraftment Table in a Local \code{dm}
}
