% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_extract_legacy.R
\name{dm_extract_legacy}
\alias{dm_extract_legacy}
\title{Select and Convert Table + Columns From Remote Source}
\usage{
dm_extract_legacy(dm_remote = dm_sql_server(), collect = TRUE, reset = FALSE)
}
\arguments{
\item{dm_remote}{\verb{[dm]} A \code{dm} object connected to the SQL Server for MLinHCT}

\item{collect}{\verb{[lgl(1)]} Should tables be collected locally on output?}

\item{reset}{\verb{[lgl(1)]} Should the cache be reset to the current results,
even if inputs have not changed? This is useful if data processing logic
has changed, but the underlying data have not.}
}
\value{
\verb{[dm]} The updated \code{dm} object
}
\description{
\code{dm_extract_legacy()} is a previous, less extensible version of \code{dm_extract()}.
It selects tables and columns of potential interest. It combines all
Cerner tables into one and joins HLA tables.
}
