% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_gvhd.R
\name{dm_gvhd_extract}
\alias{dm_gvhd_extract}
\title{Extract GVHD Table from SQL Server}
\usage{
dm_gvhd_extract(dm_remote)
}
\arguments{
\item{dm_remote}{\verb{[dm]} Remote \code{dm} connected to a SQL Server w/ HCT data}
}
\value{
\verb{[dm]} The \code{dm} w/ instructions to update the \code{gvhd} table
}
\description{
Extract GVHD Table from SQL Server
}
