% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_gvhd.R
\name{dm_gvhd_transform}
\alias{dm_gvhd_transform}
\title{Transform the GVHD Table in a Local \code{dm}}
\usage{
dm_gvhd_transform(dm_local)
}
\arguments{
\item{dm_local}{\verb{[dm]} Local \code{dm} w/ HCT data}
}
\value{
\verb{[dm]} The \code{dm} object w/ transformed \code{gvhd} table
}
\description{
Transform the GVHD Table in a Local \code{dm}
}
