% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_hct.R
\name{dm_hct}
\alias{dm_hct}
\title{Extract, Standardize, and Combine Tables from the MLinHCT Database}
\usage{
dm_hct(dm_remote = dm_sql_server(), ..., .excl_dsmb = TRUE, .quiet = FALSE)
}
\arguments{
\item{dm_remote}{\verb{[dm]} A \code{dm} object connected to the remote SQL server
database}

\item{...}{Names of tables to select; if provided, only these tables will be
extracted}

\item{.excl_dsmb}{\verb{[lgl]} Should patients marked as part of an active DSMB-monitored
study be excluded from the dataset? The default is \code{TRUE}; this should only
be changed if you are sure that your use case has been approved by the St.
Jude DSMB.}

\item{.quiet}{Should status messages be suppressed?}
}
\value{
The prepared \code{dm} object
}
\description{
\code{dm_hct()} chains together \code{dm_extract()}, \code{dm_standardize()}, and
\code{dm_combine()} to provide a single wrapper function for data preparation.
}
