% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_hla.R
\name{dm_hla_transform}
\alias{dm_hla_transform}
\title{Transform an HLA Table in a Local \code{dm}}
\usage{
dm_hla_transform(dm_local)
}
\arguments{
\item{dm_local}{Local \code{dm} with HCT data}
}
\value{
The \code{dm} with transformed \code{hla} table
}
\description{
Transform an HLA Table in a Local \code{dm}
}
