% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dm_is_remote}
\alias{dm_is_remote}
\title{Check Whether a \code{dm} Object is Connected to a Remote Server}
\usage{
dm_is_remote(dm)
}
\arguments{
\item{dm}{The \code{dm} object to check}
}
\value{
A \code{logical} indicating whether the \code{dm} is remote or not
}
\description{
Check Whether a \code{dm} Object is Connected to a Remote Server
}
\keyword{internal}
