% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_master.R
\name{dm_master_extract}
\alias{dm_master_extract}
\title{Extract Master Table from SQL Server}
\usage{
dm_master_extract(dm_remote)
}
\arguments{
\item{dm_remote}{\verb{[dm]} Remote \code{dm} connected to a SQL Server w/ HLA data}
}
\value{
The \code{dm} with instructions to update the \code{master} table
}
\description{
Extract Master Table from SQL Server
}
