% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_master.R
\name{dm_master_transform}
\alias{dm_master_transform}
\title{Transform the Master Table in a Local \code{dm}}
\usage{
dm_master_transform(dm_local)
}
\arguments{
\item{dm_local}{\verb{[dm]} Local \code{dm} with HLA data}
}
\value{
The transformed \code{dm}
}
\description{
Transform the Master Table in a Local \code{dm}
}
