% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_mrd.R
\name{dm_mrd_extract}
\alias{dm_mrd_extract}
\title{Extract MRD Table from SQL Server}
\usage{
dm_mrd_extract(dm_remote)
}
\arguments{
\item{dm_remote}{\verb{[dm]} Remote \code{dm} connected to SQL Server w/ HCT data}
}
\value{
\verb{[dm]} The \code{dm} with instructions for updating the \code{mrd} table
}
\description{
Extract MRD Table from SQL Server
}
