% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_mrd.R
\name{dm_mrd_transform}
\alias{dm_mrd_transform}
\title{Transform the MRD Table in a Local \code{dm}}
\usage{
dm_mrd_transform(dm_local)
}
\arguments{
\item{dm_local}{\verb{[dm]} Local \code{dm} with HLA data}
}
\value{
The transformed \code{dm}
}
\description{
Transform the MRD Table in a Local \code{dm}
}
