% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_relapse.R
\name{dm_relapse_clean}
\alias{dm_relapse_clean}
\title{Clean the Relapse Table in a Local \code{dm}}
\usage{
dm_relapse_clean(dm_local)
}
\arguments{
\item{dm_local}{\verb{[dm]} Local \code{dm} with HCT data}
}
\value{
\verb{[dm]} The \code{dm} object w/ transformed \code{relapse} table
}
\description{
Cleans contents of columns in the relapse table (currently does nothing).
}
