% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_relapse.R
\name{dm_relapse_extract}
\alias{dm_relapse_extract}
\title{Extract Relapse Table from SQL Server}
\usage{
dm_relapse_extract(dm_remote)
}
\arguments{
\item{dm_remote}{\verb{[dm]} Remote \code{dm} connected to SQL Server w/ HLA data}
}
\value{
\verb{[dm]} The \code{dm} object w/ instructions for updating \code{relapse} table
}
\description{
Extract Relapse Table from SQL Server
}
