% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_relapse.R
\name{dm_relapse_filter}
\alias{dm_relapse_filter}
\title{Removes Unidentifiable and Duplicate Observations in the Relapse Table}
\usage{
dm_relapse_filter(dm_local, quiet = TRUE)
}
\arguments{
\item{dm_local}{\verb{[dm]} Local \code{dm} with HCT data}

\item{quiet}{\verb{[lgl]} Whether filtering should be done silently; if \code{FALSE},
will show summary message of filtering results.}
}
\value{
\verb{[dm]} The \code{dm} object w/ filtered \code{relapse} table
}
\description{
Removes Unidentifiable and Duplicate Observations in the Relapse Table
}
