% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_relapse.R
\name{dm_relapse_std}
\alias{dm_relapse_std}
\title{Standardize Relapse Table Column Types}
\usage{
dm_relapse_std(dm_local)
}
\arguments{
\item{dm_local}{\verb{[dm]} Local \code{dm} with HCT data}
}
\value{
\verb{[dm]} The \code{dm} object with standardized Relapse table
}
\description{
Ensures that all columns are of the expected \code{class}, that all \code{character}
variables are standardized, and that all missings encoded as non-\code{NA} values
are recoded to \code{NA}.
}
