% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_relapse.R
\name{dm_relapse_transform}
\alias{dm_relapse_transform}
\title{Transform the Relapse Table in a Local \code{dm}}
\usage{
dm_relapse_transform(dm_local)
}
\arguments{
\item{dm_local}{\verb{[dm]} Local \code{dm} with HCT data}
}
\value{
\verb{[dm]} The \code{dm} object w/ transformed \code{relapse} table
}
\description{
Transform the Relapse Table in a Local \code{dm}
}
