% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-dm.R
\name{dm_sort}
\alias{dm_sort}
\title{Sort a \code{dm} Object by Table Name}
\usage{
dm_sort(dm)
}
\arguments{
\item{dm}{A \code{dm} object to sort}
}
\value{
The sorted \code{dm}
}
\description{
\code{dm_sort()} converts a \code{dm} to a list, sorts the list, and converts back to
a \code{dm}. Currently, all \code{dm}-specific properties (specifically primary and
foreign keys) are lost during this process.
}
\keyword{internal}
