% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy_dm_transform.R
\name{dm_transform}
\alias{dm_transform}
\title{Transform Tables to Analysis-Friendly Format}
\usage{
dm_transform(dm_local = dm_extract_legacy(), reset = FALSE)
}
\arguments{
\item{dm_local}{\verb{[dm]} A local \code{dm} object from \code{dm_extract()}}

\item{reset}{\verb{[lgl(1)]} Should the cache be reset to the current results,
even if inputs have not changed? This is useful if data processing logic
has changed, but the underlying data have not.}
}
\value{
\verb{[dm]} The updated \code{dm} object
}
\description{
A previous version of the dmhct pipeline performed all transformations of
tables simultaneously; to ensure backwards compatibility, this behavior has
been retained in \code{dm_transform()}. However, it is strongly recommended that
new code not use \code{dm_transform()} and instead use the updated pipeline.
}
