% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-export.R
\name{intvl_to_matrix}
\alias{intvl_to_matrix}
\title{Convert Standardized Intervals to Matrix Format}
\usage{
intvl_to_matrix(x)
}
\arguments{
\item{x}{A \code{character} vector of standardized intervals}
}
\value{
A 4-column numeric matrix:
\itemize{
\item \code{left_closed}:  Whether the left bound is closed or open
\item \code{left_bound}:   The left bound of the interval
\item \code{right_bound}:  The right bound of the interval
\item \code{right_closed}: Whether the right bound is closed or open
}
}
\description{
\code{intvl_to_matrix()} converts interval representation standardized by
\code{std_invl()} to a 4-column \code{numeric} matrix. Columns represent open or
closed bounds and the location of those bounds.
}
