% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_patterns.R
\docType{data}
\name{na_patterns}
\alias{na_patterns}
\title{Common Patterns Representing Missing Data}
\format{
An object of class \code{character} of length 8.
}
\usage{
na_patterns
}
\description{
\code{na_patterns} is a collection of regular expression that commonly represent
missing data, especially when the character vector should be converted to
something else. These are designed to match strings that have already been
standardized.
}
\keyword{datasets}
