% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-export.R
\name{non_numeric}
\alias{non_numeric}
\title{Extract Values That Cannot Be Converted to \code{numeric}}
\usage{
non_numeric(x, unique = TRUE, sort = unique, std_num = FALSE)
}
\arguments{
\item{x}{A vector}

\item{unique}{Whether unique values should be returned; if \code{FALSE}, all
values are returned}

\item{sort}{Whether return values should be sorted; most useful when
\code{unique = TRUE}}

\item{std_num}{Whether to use \code{std_num()} for numeric conversion; if \code{FALSE},
conversion is performed directly by \code{as.numeric()} (with warnings
suppressed)}
}
\value{
The values of \code{x} that resulted in \code{NA_real_} after conversion; this
includes any \code{NA} values in \code{x} before conversion
}
\description{
\code{non_numeric()} is designed primarily for interactive checking of numeric
conversions. It helps quickly determine what values in a vector cannot be
converted to \code{numeric} (either directly or via \code{std_num()}); this is
particularly useful for checking steps of a data cleaning pipeline.
}
