% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{path_create}
\alias{path_create}
\title{Create a Path from Components}
\usage{
path_create(...)
}
\arguments{
\item{...}{\verb{[chr]} Components of a path}
}
\value{
\verb{[fs_path]} The created and cleaned file path
}
\description{
Create a Path from Components
}
\keyword{internal}
