% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_cerner.R
\name{std_cerner_tbl}
\alias{std_cerner_tbl}
\title{Standardize a Cerner Table}
\usage{
std_cerner_tbl(
  tbl,
  path_var = system.file("extdata", "cerner_var_map.csv", package = "predHCT")
)
}
\arguments{
\item{tbl}{\verb{[tbl_dbi]} A table containing Cerner data}
}
\value{
A \code{tbl_dbi} with standardized columns
}
\description{
Selects needed columns from a Cerner table and converts them to the expected
class. Test names are replaced with standardized variable names and rows
with missing information are dropped.
}
\keyword{internal}
