% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-export.R
\name{std_chr}
\alias{std_chr}
\title{Standardize \code{character} Vectors}
\usage{
std_chr(
  x,
  case = c("upper", "lower", "title", "sentence"),
  keep_inner_newlines = TRUE
)
}
\arguments{
\item{x}{A character vector}

\item{case}{The case to convert to. \code{NULL} will skip case conversion.}

\item{keep_inner_newlines}{Whether to retain line breaks inside text. \code{FALSE}
will treat newlines and carriage returns identically to any other whitespace.}
}
\value{
The standardized character vector
}
\description{
\code{std_chr()} standardizes \code{character} vectors to ASCII text with no unnecessary
whitespace and a given case. By default, it will retain newlines inside text,
though it will condense consecutive newlines and any carriage returns into a
single newline.
}
