% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_hla.R
\name{std_hla_allele}
\alias{std_hla_allele}
\title{Standardize HLA Allele Representations}
\usage{
std_hla_allele(x)
}
\arguments{
\item{x}{\verb{[character]} A vector of allele IDs}
}
\value{
Standardized alleles
}
\description{
Standardize HLA Allele Representations
}
\keyword{internal}
