% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_hla.R
\name{std_hla_tbl}
\alias{std_hla_tbl}
\title{Standardize an HLA Table}
\usage{
std_hla_tbl(tbl, na = c("", "NT", "Blank", "-", "Not Interpretable"))
}
\arguments{
\item{tbl}{\verb{[data.frame]} HLA data}

\item{na}{\verb{[character]} Vector of values to interpret as \code{NA}}
}
\value{
The standardized data
}
\description{
Standardize an HLA Table
}
\keyword{internal}
