% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{std_intvl}
\alias{std_intvl}
\title{Standardize Interval Representations}
\usage{
std_intvl(x, std_chr = TRUE, warn = TRUE, chimerism = c("no", "donor", "host"))
}
\arguments{
\item{x}{A \code{character} vector}

\item{std_chr}{Whether to standarize the strings before parsing}

\item{warn}{Whether to emit a warning when potential numeric values are not
able to be converted to an interval}

\item{chimerism}{The type of chimerism to extract, if any; passed to \code{chr_to_num()},
when \code{std_chr = TRUE}, ignored otherwise.}
}
\value{
A \code{character} vector
}
\description{
\code{std_invl()} standardizes the various representations of numeric intervals
found in the ML in HCT dataset. These intervals are assumed to be in percentage
values and thus lie between 0 and 100. Explicit intervals with upper and lower
bounds, as well as implicit intervals using < and >, are handled (<= and >=
are currently not supported). The return value simplifies to </>/<=/>= or
a single numeric value if possible and uses standard interval notation if not.
}
