% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-export.R
\name{std_lgl}
\alias{std_lgl}
\title{Standardize \code{logical} Representations in Various Formats}
\usage{
std_lgl(
  x,
  true = c("^TRUE$", "^1$", "^YES", "^POS", "^ALIVE", "^ON THERAPY"),
  false = c("^FALSE$", "^0$", "^NO", "^NEG", "^EXPIRED", "^DECEASED", "^OFF THERAPY"),
  na = na_patterns,
  std_chr = TRUE,
  warn = TRUE
)
}
\arguments{
\item{x}{A vector to convert}

\item{true}{Regex patterns to consider \code{TRUE}. Passed to
\code{stringr::str_detect()}. Can be a vector of patterns.}

\item{false}{Regex patterns to consider \code{FALSE}. Passed to
\code{stringr::str_detect()}. Can be a vector of patterns.}

\item{na}{Regex patterns to consider \code{NA}. Passed to
\code{stringr::str_detect()}. Can be a vector of patterns.}

\item{std_chr}{Whether to standardized a \code{character} vector before parsing}

\item{warn}{Whether to warn if \code{character} strings were not converted to
\code{logical}}
}
\value{
A \code{logical} vector
}
\description{
\code{std_lgl()} converts other classes to \code{logical} vectors. All but \code{character}
use \code{as.logical()}; \code{character} vectors are converted by first (optionally)
standardizing with \code{std_chr} and then assigning logical value based on the
regular expression in \code{true}, \code{false}, and \code{na}.
}
