% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{std_num}
\alias{std_num}
\title{Convret and Standardize Numeric Values in Various Forms}
\usage{
std_num(x, std_chr = TRUE, warn = TRUE)
}
\arguments{
\item{x}{A vector to convert to numeric}

\item{std_chr}{Whether to standardize a \code{character} or \code{factor} before conversion}

\item{warn}{Whether to warn when strings cannot be converted; passed to \code{chr_to_num()}}
}
\value{
A \code{numeric} vector
}
\description{
\code{std_num()} converts all base classes, as well as \code{int64}, \code{factor}, \code{Date},
and \code{POSIXt} vectors to the simplest numeric form possible.
}
\details{
\code{character} vectors are standardized using \code{std_chr()} by default, then
converted. \code{factor}s are treated as \code{character} vectors, rather than using
the underlying integer representation. \code{double} and \code{int64} vectors will be
converted to \code{integer} if this does not cause overflow or loss of precision.
\code{Date} is converted to \code{integer}, and \code{POSIXt} is converted to \code{integer} if
the range allows, otherwise \code{double}.
}
