% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{str_to_na}
\alias{str_to_na}
\title{Replace Detected Patterns with \code{NA}}
\usage{
str_to_na(string, pattern = stringr::fixed(""))
}
\arguments{
\item{string}{Input vector. Either a character vector, or something
coercible to one.}

\item{pattern}{Pattern to look for.

The default interpretation is a regular expression, as described in
\code{vignette("regular-expressions")}. Use \code{\link[stringr:regex]{regex()}} for finer control of the
matching behaviour.

Match a fixed string (i.e. by comparing only bytes), using
\code{\link[stringr:fixed]{fixed()}}. This is fast, but approximate. Generally,
for matching human text, you'll want \code{\link[stringr:coll]{coll()}} which
respects character matching rules for the specified locale.

Match character, word, line and sentence boundaries with
\code{\link[stringr:boundary]{boundary()}}. An empty pattern, "", is equivalent to
\code{boundary("character")}.}
}
\value{
The input as a character vector with \code{pattern}s replaced with \code{NA}
}
\description{
\code{str_to_na()} detects \code{pattern}s in \code{x} using
\code{\link[stringr:str_detect]{stringr::str_detect()}} and replaces all
indices where a \code{pattern} is found with \code{NA_character_}. It is vectorized
over both \code{x} and \code{pattern}; \code{pattern} will be looped over if multipel are
provided.
}
\keyword{internal}
