% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_cache.R
\name{tbl_checksum_local}
\alias{tbl_checksum_local}
\title{Hash a Local Table}
\usage{
tbl_checksum_local(tbl_local)
}
\arguments{
\item{tbl_local}{A local table in a \code{dm} object}
}
\value{
The checksum of the table
}
\description{
Hash a Local Table
}
