% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_cerner.R
\docType{data}
\name{UtilsCerner}
\alias{UtilsCerner}
\alias{utils_cerner}
\title{Utility Functions for Cerner Table ELT}
\description{
Collection of utility functions for cerner data
}
\keyword{Cerner}
\keyword{Standardize}
\keyword{Table}
\keyword{a}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UtilsCerner-std_cerner_tbl}{\code{UtilsCerner$std_cerner_tbl()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UtilsCerner-std_cerner_tbl"></a>}}
\if{latex}{\out{\hypertarget{method-UtilsCerner-std_cerner_tbl}{}}}
\subsection{Method \code{std_cerner_tbl()}}{
Selects needed columns from a Cerner table and converts them to the
expected class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UtilsCerner$std_cerner_tbl(tbl)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tbl}}{\verb{[tbl_dbi]} A table containing Cerner data}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{tbl_dbi} with standardized columns
}
}
}
