% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_chimerism.R
\docType{data}
\name{UtilsChimerism}
\alias{UtilsChimerism}
\alias{utils_chimerism}
\title{Utility Functions for Chimerism Table ELT}
\description{
Collection of utility functions for chimerism data
}
\keyword{Chimerism}
\keyword{Data}
\keyword{Detect}
\keyword{Factor}
\keyword{Level}
\keyword{in}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UtilsChimerism-str_detect_fct}{\code{UtilsChimerism$str_detect_fct()}}
\item \href{#method-UtilsChimerism-str_range_to_num}{\code{UtilsChimerism$str_range_to_num()}}
\item \href{#method-UtilsChimerism-str_excel_dt_to_num}{\code{UtilsChimerism$str_excel_dt_to_num()}}
\item \href{#method-UtilsChimerism-normalize}{\code{UtilsChimerism$normalize()}}
\item \href{#method-UtilsChimerism-probit}{\code{UtilsChimerism$probit()}}
\item \href{#method-UtilsChimerism-inv_probit}{\code{UtilsChimerism$inv_probit()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UtilsChimerism-str_detect_fct"></a>}}
\if{latex}{\out{\hypertarget{method-UtilsChimerism-str_detect_fct}{}}}
\subsection{Method \code{str_detect_fct()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UtilsChimerism$str_detect_fct(x, lvl, lbl)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{\verb{[chr]} A character vector}

\item{\code{lvl}}{\verb{[chr(1)]} The level's numeric representation}

\item{\code{lbl}}{\verb{[chr(1)]} The level's label}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\verb{[lgl(1)]} Presence or absence of the indicated factor level
Convert a \% Range to It's Midpoint
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UtilsChimerism-str_range_to_num"></a>}}
\if{latex}{\out{\hypertarget{method-UtilsChimerism-str_range_to_num}{}}}
\subsection{Method \code{str_range_to_num()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UtilsChimerism$str_range_to_num(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{\verb{[chr]} Character representation of a range of percentages}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\verb{[dbl]} The midpoint of each range of values
Convert Excel Date to Number

Converts Excel dates in MDY format to a number using an origin of
"1/0/1900". Elements that cannot be converted are returned as-is.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UtilsChimerism-str_excel_dt_to_num"></a>}}
\if{latex}{\out{\hypertarget{method-UtilsChimerism-str_excel_dt_to_num}{}}}
\subsection{Method \code{str_excel_dt_to_num()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UtilsChimerism$str_excel_dt_to_num(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{\verb{[chr]} Character vector containing Excel dates}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\verb{[chr]} Numeric representation of Excel date, where available
Normalize a \code{numeric} Vector to \verb{[0, 1]}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UtilsChimerism-normalize"></a>}}
\if{latex}{\out{\hypertarget{method-UtilsChimerism-normalize}{}}}
\subsection{Method \code{normalize()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UtilsChimerism$normalize(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{\verb{[dbl]} A \code{numeric} vector or a vector than can be coerced
to \code{numeric}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\verb{[dbl]} The normalized vector
Transform a \code{numeric} Vector Using the Probit Function
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UtilsChimerism-probit"></a>}}
\if{latex}{\out{\hypertarget{method-UtilsChimerism-probit}{}}}
\subsection{Method \code{probit()}}{
Transforms a \code{numeric} vector using the Gaussian quantile function
\code{\link[stats:Normal]{qnorm()}}. Converts the vector to the
\verb{[0, 1]} range using \code{range} prior to transformation. Since \code{qnorm(0)}
and \code{qnorm(1)} are infinite, values closer than \code{tol} to \code{0} or \code{1} are
truncated before transformation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UtilsChimerism$probit(x, range = c(0, 100), tol = sqrt(.Machine$double.eps))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{\verb{[dbl]} A vector to transform}

\item{\code{range}}{\verb{[dbl(2)]} The lower and upper bounds of the domain of \code{x}}

\item{\code{tol}}{\verb{[dbl(1)]} The tolerance for truncating \code{x} on the \verb{[0, 1]}
scale; must be between \code{0} and \code{0.01}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\verb{[dbl]} The transformed vector
Transform a \code{numeric} Vector Using the Inverse Probit Function
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UtilsChimerism-inv_probit"></a>}}
\if{latex}{\out{\hypertarget{method-UtilsChimerism-inv_probit}{}}}
\subsection{Method \code{inv_probit()}}{
Transforms a \code{numeric} vector using the Gaussian distribution function
\code{\link[stats:Normal]{pnorm()}}. As the inverse of \code{probit()}, also
inverts truncation by replacing transformed values closer than \code{tol} to
\code{0} or \code{1} with \code{0} or \code{1} (respectively). Maps result to the domain
specified by \code{range}; values outside of \code{range} will be truncated.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UtilsChimerism$inv_probit(
  x,
  range = c(0, 100),
  tol = sqrt(.Machine$double.eps)
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{\verb{[dbl]} A vector to transform}

\item{\code{range}}{\verb{[dbl(2)]} The lower and upper bounds of the range of the result}

\item{\code{tol}}{\verb{[dbl(1)]} The tolerance used for truncating in the \code{probit} transform}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\verb{[dbl]} The transformed vector
}
}
}
