% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_hla.R
\docType{data}
\name{UtilsHLA}
\alias{UtilsHLA}
\alias{utils_hla}
\title{Utility Functions for HLA Table ELT}
\description{
Collection of utility functions for HLA data
}
\keyword{Allele}
\keyword{HLA}
\keyword{Representations}
\keyword{Standardize}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UtilsHLA-std_hla_allele}{\code{UtilsHLA$std_hla_allele()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UtilsHLA-std_hla_allele"></a>}}
\if{latex}{\out{\hypertarget{method-UtilsHLA-std_hla_allele}{}}}
\subsection{Method \code{std_hla_allele()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UtilsHLA$std_hla_allele(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{\verb{[chr]} A vector of allele IDs}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\verb{[chr]} Standardized alleles
}
}
}
