% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_master.R
\docType{data}
\name{UtilsMaster}
\alias{UtilsMaster}
\alias{utils_master}
\title{Utility Functions for Master Table ELT}
\description{
Collection of utility functions for master data
}
\keyword{Convert}
\keyword{Count}
\keyword{Data}
\keyword{Degree}
\keyword{Denominator}
\keyword{Match}
\keyword{Specified}
\keyword{`character`}
\keyword{internal}
\keyword{of}
\keyword{to}
\keyword{with}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UtilsMaster-convert_degree_match}{\code{UtilsMaster$convert_degree_match()}}
\item \href{#method-UtilsMaster-clone}{\code{UtilsMaster$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UtilsMaster-convert_degree_match"></a>}}
\if{latex}{\out{\hypertarget{method-UtilsMaster-convert_degree_match}{}}}
\subsection{Method \code{convert_degree_match()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UtilsMaster$convert_degree_match(x, denom = 6L)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{\verb{[character]} Vector of degree of match counts}

\item{\code{denom}}{\verb{[integer(1)]} Denominator of counts to extract}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An \code{integer} vector of counts
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UtilsMaster-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UtilsMaster-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UtilsMaster$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
