% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_relapse.R
\name{UtilsRelapse}
\alias{UtilsRelapse}
\alias{utils_relapse}
\title{Utility Functions for MRD Table ELT}
\description{
Collection of utility functions for MRD data
}
\keyword{"`).}
\keyword{(currently}
\keyword{(i.e.}
\keyword{Missing}
\keyword{Parenthetic}
\keyword{Standardize}
\keyword{`";}
\keyword{`"NA"`,}
\keyword{`"Other"`,}
\keyword{`mcat`}
\keyword{`std_mcat()`}
\keyword{a}
\keyword{amount}
\keyword{and}
\keyword{are}
\keyword{be}
\keyword{can}
\keyword{categorical}
\keyword{categories}
\keyword{category.}
\keyword{data}
\keyword{delimiter}
\keyword{ensure}
\keyword{etc.)}
\keyword{expressions}
\keyword{form}
\keyword{in}
\keyword{internal}
\keyword{into}
\keyword{is}
\keyword{minimal}
\keyword{multi-membership}
\keyword{necessary}
\keyword{not}
\keyword{of}
\keyword{own}
\keyword{performs}
\keyword{relapse}
\keyword{removed}
\keyword{single}
\keyword{split}
\keyword{split.}
\keyword{standardization}
\keyword{that}
\keyword{the}
\keyword{their}
\keyword{they}
\keyword{this}
\keyword{to}
\keyword{uninformative}
\keyword{values}
\keyword{variables}
\keyword{via}
\keyword{when}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UtilsRelapse-std_mcat}{\code{UtilsRelapse$std_mcat()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UtilsRelapse-std_mcat"></a>}}
\if{latex}{\out{\hypertarget{method-UtilsRelapse-std_mcat}{}}}
\subsection{Method \code{std_mcat()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UtilsRelapse$std_mcat(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{\verb{[chr]} A character vector}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\verb{[chr]} The standardized character vector
factor level
}
}
}
