% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{chr_to_num}
\alias{chr_to_num}
\title{Clean and Possibly Convert Numbers in \code{character} Representation}
\usage{
chr_to_num(
  x,
  std = TRUE,
  warn = TRUE,
  convert = TRUE,
  na_patterns = c("^-*$", "(?:^|\\\\W)N/?A(?:$|\\\\W)",
    "NOT? (?:RER?PORT|(?:[A-Z ]|-)*PHENOTYPE)",
    "NOT? (?:AMP|AVAIL|DATA|DONE|EVAL|FOLLOWED|IDENT|VALUE)",
    "(?:^|\\\\W)(?:NG|UNK|CQNS|QNS|QIS|NSQ|IQ|QI|NR)(?:$|\\\\W)",
    "(?:^|\\\\W)(?:UNKNOWN|INCONCLUSIVE|INSUFFICIENT)(?:$|\\\\W)", "-999[0-9]",
    "^REPORTED$"),
  chimerism = c("no", "donor", "host")
)
}
\arguments{
\item{x}{A \code{character} vector}

\item{std}{Whether to standardize the vector before cleaning and converting}

\item{convert}{Whether to actually convert to \code{numeric}}

\item{na_patterns}{Regular expressions to convert to \code{NA}}

\item{chimerism}{The type of chimerism to extract, if any}
}
\value{
A \code{numeric} or \code{character} vector
}
\description{
\code{chr_to_num()} cleans strings representing numeric values. It is tailored for
use with the ML in HCT dataset; use with other data may not go as expected.
}
\details{
The function first converts strings matching \code{na_patterns} to missing values.
It then simplifies any numeric representations it finds, including 10^x and
various common typos observed in the data (unneeded decimals, commas, and zeros).
Additionally, it converts Excel datetimes of the form 1/x/1900 hh:mm back to
decimal representation, and additionally converts fractions less than 1 to
decimals. Lastly, it handles a few idiosyncratic text strings, including
conversion of POSITIVE and NEGATIVE values, as well as W(E|A)KLY POSITIVE
and IN CR (complete remission). Optionally, it will extract values labelled as
donor or host (patient), which is specific to the chimerism dataset. It also
removes the leading text MRD BY NGS. By default, the function emits a warning
when potential numeric values are not able to be converted to \code{numeric}.
}
\keyword{internal}
