% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_chimerism.R
\name{dm_chimerism_transform}
\alias{dm_chimerism_transform}
\title{Transform the Chimerism Table in a Local \code{dm}}
\usage{
dm_chimerism_transform(dm_local)
}
\arguments{
\item{dm_local}{\verb{[dm]} Local \code{dm} w/ HCT data}
}
\value{
\verb{[dm]} The \code{dm} object w/ transformed \code{chimerism} table
}
\description{
Transform the Chimerism Table in a Local \code{dm}
}
