% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-dm.R
\name{dm_collect}
\alias{dm_collect}
\title{Collect All Tables in a \code{dm} Object}
\usage{
dm_collect(dm_remote)
}
\arguments{
\item{dm_remote}{\verb{[dm]} A \code{dm} object connected to a remote source}
}
\value{
\verb{[dm]} A new \code{dm} object containing the collected (local) tables
}
\description{
Collect All Tables in a \code{dm} Object
}
