% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-dm.R
\name{dm_compute}
\alias{dm_compute}
\title{Compute All Tables in a \code{dm} Object}
\usage{
dm_compute(dm_remote, quiet = TRUE)
}
\arguments{
\item{dm_remote}{\verb{[dm]} A \code{dm} object connected to a remote source}
}
\value{
\verb{[dm]} The updated object with tables computed
}
\description{
Compute All Tables in a \code{dm} Object
}
