% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_disease_status.R
\name{dm_disease_status_extract}
\alias{dm_disease_status_extract}
\title{Extract Disease Status Table from SQL Server}
\usage{
dm_disease_status_extract(dm_remote)
}
\arguments{
\item{dm_remote}{Remote \code{dm} object connected to SQL Server}
}
\value{
\code{dm} object with instructions for updating \code{disease_status} table
}
\description{
Extract Disease Status Table from SQL Server
}
