% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_elt.R
\name{dm_elt}
\alias{dm_elt}
\title{Extract, Load, and Transform Remote Tables to Local Source}
\usage{
dm_elt(dm_remote = dm_sql_server(), reset = FALSE, close = NULL)
}
\arguments{
\item{dm_remote}{\verb{[dm]} Remote \code{dm} object containing HCT data}

\item{reset}{\verb{[lgl(1)]} Should the cache be reset to the current results,
even if inputs have not changed? This is useful if data processing logic
has changed, but the underlying data have not.}

\item{close}{\verb{[lgl(1)]} Whether to close the SQL Server connection on
exit. \code{NULL} closes if \code{dm_remote} has attribute \code{default == TRUE} and
leaves open otherwise.}

\item{cache}{\verb{[lgl(1)]} Should results be cached (if inputs have changed)
or read from cache (if inputs have not changed)?}
}
\value{
\verb{[dm]} A \code{dm} object
}
\description{
Extract, Load, and Transform Remote Tables to Local Source
}
