% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_elt.R
\name{dm_elt}
\alias{dm_elt}
\title{Extract, Load, and Transform Remote Tables to Local Source}
\usage{
dm_elt(dm_remote = sql_server_dm(), cache = TRUE, reset = FALSE, close = NULL)
}
\arguments{
\item{dm_remote}{Remote \code{dm} object containing HCT data}

\item{cache}{Whether to use the cache. If \code{cache == TRUE} and \code{reset == TRUE},
the data will be ELT'd from the source and cached, even if an up-to-date
cache exists.}

\item{reset}{Whether to force collection and caching of data.}

\item{close}{Whether to close the SQL Server connection on exit. \code{NULL}
closes if \code{dm_remote} has attribute \code{default == TRUE} and leaves open
otherwise.}
}
\value{
A \code{dm} object
}
\description{
Extract, Load, and Transform Remote Tables to Local Source
}
