% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_engraftment.R
\name{dm_engraftment_transform}
\alias{dm_engraftment_transform}
\title{Transform the Engraftment Table in a Local \code{dm}}
\usage{
dm_engraftment_transform(dm_local)
}
\arguments{
\item{dm_local}{\verb{[dm]} Local \code{dm} w/ HCT data}
}
\value{
\verb{[dm]} The \code{dm} object w/ transformed \code{engraftment} table
}
\description{
Transform the Engraftment Table in a Local \code{dm}
}
