% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_elt.R
\name{dm_extract}
\alias{dm_extract}
\title{Select and Convert Table + Columns From Remote Source}
\usage{
dm_extract(
  dm_remote = sql_server_dm(),
  collect = TRUE,
  compute = FALSE,
  cache = collect,
  reset = FALSE
)
}
\arguments{
\item{dm_remote}{\verb{[dm]} A \code{dm} object connected to the SQL Server for MLinHCT}
}
\value{
The updated \code{dm} object
}
\description{
\code{dm_extract()} selects tables and columns of potential interest. It
combines all Cerner tables into one and joins HLA tables.
}
