% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_extract.R
\name{dm_extract}
\alias{dm_extract}
\title{Extract Remote Tables from SQL Server for MLinHCT}
\usage{
dm_extract(
  dm_remote = dm_sql_server(),
  ...,
  .collect = TRUE,
  .legacy = FALSE,
  .reset = FALSE,
  .excl_dsmb = FALSE,
  .quiet = FALSE,
  reset = .reset
)
}
\arguments{
\item{dm_remote}{\verb{[dm]} A \code{dm} object connected to the remote SQL server
database}

\item{...}{Names of tables to select; if provided, only these tables will be
extracted}

\item{.collect}{\verb{[lgl]} Indicates whether the extracted data should be loaded
onto the local machine (\code{TRUE} by default)}

\item{.legacy}{\verb{[lgl]} Should the legacy version of \code{dm_extract()} be used?
Will be deprecated in a future release, along with \code{.reset}.}

\item{.reset}{\verb{[lgl]} Should the legacy cache be forced to reset? Only applicable
if \code{.legacy = TRUE}; ignored otherwise. Will be deprecated in a future release,
along with \code{.legacy}.}

\item{.excl_dsmb}{\verb{[lgl]} \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} This information
is no longer available in the remote database.}

\item{.quiet}{Should status messages be suppressed?}

\item{reset}{\verb{[lgl]} \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Please use
\code{.reset} instead. Current behavior will only consider this argument if
\code{.reset} is unchanged from the default.}
}
\value{
\verb{[dm]} A \code{dm} object with all tables and columns extracted from the
remote source.
}
\description{
\code{dm_extract()} extracts and (optionally) loads the remote database housing
the MLinHCT into the current R session. Unless \code{.legacy = TRUE}, column and
table names are standardized during extraction, but no other operations are
performed. When \code{.legacy = TRUE}, the legacy version of \code{dm_extract()} is
used; see details for this behavior. Note that legacy behavior will be
deprecated and eventually removed in future releases, so it is strongly
recommended that any new code use \code{.legacy = FALSE}.
}
\details{
Legacy behavior is more opinionated than the current version of \code{dm_extract()}.
First, only a subset of tables and columns are extracted. Second, HLA tables
and Cerner tables are combined into a single HLA table and a single Cerner
table. Third, some column standardization occurs (though it is limited to
simple \code{as()} transformations, \code{trimws(toupper(x))} on character variables,
and replacement of implicit missing values with explicit \code{NA}s.) Finally,
some filtering of "uninformative" observations may occur. In the current
pipeline, these changes are deferred to later steps to give more control
to the user.
}
