% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_hla.R
\name{dm_hla_extract}
\alias{dm_hla_extract}
\title{Extract and Unite HLA Tables in SQL Server}
\usage{
dm_hla_extract(dm_remote, quiet = TRUE)
}
\arguments{
\item{dm_remote}{\verb{[dm]} Remote \code{dm} connected to a SQL Server w/ HCT data}

\item{quiet}{\verb{[lgl(1)]} Should the function return quietly if an HLA table is
already present?}
}
\value{
\verb{[dm]} The \code{dm} w/ instructions to create a combined \code{hla} table
}
\description{
Extract and Unite HLA Tables in SQL Server
}
