% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_hla.R
\name{dm_hla_transform}
\alias{dm_hla_transform}
\title{Extract and Unite HLA Tables in SQL Server
Transform an HLA Table in a Local \code{dm}}
\usage{
dm_hla_transform(dm_local)
}
\arguments{
\item{dm_local}{\verb{[dm]} Local \code{dm} w/ HCT data}
}
\value{
\verb{[dm]} The \code{dm} w/ transformed \code{hla} table
}
\description{
Extract and Unite HLA Tables in SQL Server
Transform an HLA Table in a Local \code{dm}
}
