% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_relapse.R
\name{dm_relapse_transform}
\alias{dm_relapse_transform}
\title{Temporary Aggregation of Standardize/Clean/Filter Workflow}
\usage{
dm_relapse_transform(dm_local)
}
\arguments{
\item{dm_local}{\verb{[dm]} Local \code{dm} with HCT data}
}
\value{
\verb{[dm]} The \code{dm} object with transformed relapse table
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Runs \code{dm_relapse_std()}, \code{dm_relapse_clean()}, and \code{dm_relapse_filter()}
sequentially to maintain compatibility with previous API. Please use these
functions individually instead, as they provided finer-grained control over
where your data is in the pipeline.
}
\keyword{internal}
