% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_elt.R
\name{dm_transform}
\alias{dm_transform}
\title{Transform Tables to Analysis-Friendly Format}
\usage{
dm_transform(dm_local = dm_extract(), cache = TRUE, reset = FALSE)
}
\arguments{
\item{dm_local}{\verb{[dm]} A local \code{dm} object from \code{dm_extract()}}
}
\value{
\verb{[dm]} The updated \code{dm} object
}
\description{
Transform Tables to Analysis-Friendly Format
}
