% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmhct-package.R
\docType{package}
\name{dmhct-package}
\alias{dmhct}
\alias{dmhct-package}
\title{dmhct: A Data Model Package for the MLinHCT Project}
\description{
Extracts, Loads, and Transforms data from the SQL Server containing HCT data to a `dm` object with cleaned tables.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jesse-smith/dmhct}
  \item \url{https://jesse-smith.github.io/dmhct/}
  \item Report bugs at \url{https://github.com/jesse-smith/dmhct/issues}
}

}
\author{
\strong{Maintainer}: Jesse Smith \email{jesse.smith@stjude.org}

}
\keyword{internal}
