% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_cerner.R
\name{std_cerner_tbl}
\alias{std_cerner_tbl}
\title{Standardize a Cerner Table}
\usage{
std_cerner_tbl(tbl)
}
\arguments{
\item{tbl}{\verb{[tbl_dbi]} A table containing Cerner data}
}
\value{
A \code{tbl_dbi} with standardized columns
}
\description{
Selects needed columns from a Cerner table and converts them to the expected
class.
}
\keyword{internal}
