% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-export.R
\name{std_intvl}
\alias{std_intvl}
\title{Standardize Interval Representations}
\usage{
std_intvl(
  x,
  less_than = c("LESS THAN",
    "[A-Z ]*NOTHING TO SUGGEST[A-Z ]*SENSITIVITY[A-Z ]*(?=[0-9])"),
  greater_than = c("GREATER THAN"),
  na = na_patterns,
  std_chr = TRUE,
  warn = TRUE,
  ...
)
}
\arguments{
\item{x}{A \code{character} vector}

\item{less_than}{Regex patterns to consider \code{"<"}. Passed to
\code{stringr::str_replace()}. Can be a vector of patterns.}

\item{greater_than}{Regex patterns to consider \code{">"}. Passed to
\code{stringr::str_replace()}. Can be a vector of patterns.}

\item{na}{Regex patterns to consider \code{NA}. Passed to \code{stringr::str_detect()}.
Can be a vector of patterns.}

\item{std_chr}{Whether to standarize the strings before parsing}

\item{warn}{Whether to emit a warning when potential numeric values are not
able to be converted to an interval}

\item{...}{
  Arguments passed on to \code{\link[=chr_to_num]{chr_to_num}}
  \describe{
    \item{\code{std}}{Whether to standardize the vector before cleaning and converting}
    \item{\code{convert}}{Whether to actually convert to \code{numeric}}
    \item{\code{replace}}{A \code{data.frame} of regular expressions and strings to replace
them; regular expression should be in a column named \code{pattern}, and
replacements should be in a column named \code{replacement}. Each row is passed
to \code{stringr::str_replace()}.}
    \item{\code{per_action}}{How to treat \%/percent/per million/etc labels. \code{drop} simply
removes the labels, \code{divide} divides the value by the appropriate denominator,
and \code{ignore} does nothing.}
    \item{\code{multiple_decimals}}{How to handle multiple decimals within a number}
    \item{\code{donor_host}}{Which value to use when values for both a donor and a host
are given}
  }}
}
\value{
A \code{character} vector
}
\description{
\code{std_invl()} standardizes the various representations of numeric intervals
found in the ML in HCT dataset. These intervals are assumed to be in percentage
values and thus lie between 0 and 100. Explicit intervals with upper and lower
bounds, as well as implicit intervals using < and >, are handled (<= and >=
are currently not supported). The return value simplifies to </>/<=/>= or
a single numeric value if possible and uses standard interval notation if not.
}
