% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-date.R
\docType{data}
\name{UtilsDate}
\alias{UtilsDate}
\title{Date Conversion Utilities}
\format{
An object of class \code{UtilsDate} (inherits from \code{R6}) of length 3.
}
\usage{
UtilsDate
}
\arguments{
\item{force}{Whether to force conversion to \code{Date} (\code{force = "dt"}) or
\code{POSIXct} (\code{force = "dttm"}). The default is no forcing (\code{force = "none"}).}

\item{tz_heuristic}{Hours to consider in determining presence of sub-daily
information. Only exact hours (i.e. 5:00:00) will be combined. The default
corresponds to accidental encoding of the CST-UTC offset as hours.}

\item{x}{A vector of character dates}

\item{tz}{Optional timezone}

\item{orders}{The orders to use when parsing character vector with
\code{\link[lubridate:parse_date_time]{parse_date_time()}}}

\item{...}{Additional arguments to pass to
\code{\link[lubridate:parse_date_time]{parse_date_time()}}}
}
\value{
Either a \code{POSIXct} vector or a \code{Date} vector
Parse Dates in Character Format to Datetime Format

\code{chr_to_dttm} standardizes a datetime vector in character format and returns
a vector in \code{POSIXct} format. It is intended for use in \code{janitor::convert_to_datetime};
it thus muffles certain warnings that are duplicated by that function.

A \code{POSIXct} vector
}
\description{
Contains utilities for standardizing dates and datetimes
}
\keyword{(i.e.}
\keyword{Coerce}
\keyword{Dates}
\keyword{Datetimes}
\keyword{If}
\keyword{Information}
\keyword{Lost}
\keyword{No}
\keyword{Specifically,}
\keyword{`1L`),}
\keyword{`Date`}
\keyword{`POSIXt`}
\keyword{`dttm_to_dt()`}
\keyword{`dttm_to_dt`}
\keyword{a}
\keyword{additional}
\keyword{all}
\keyword{and}
\keyword{attempted}
\keyword{by}
\keyword{checks}
\keyword{coerced.}
\keyword{contained}
\keyword{conversion}
\keyword{converts}
\keyword{data}
\keyword{datetime}
\keyword{each}
\keyword{for}
\keyword{format}
\keyword{format,}
\keyword{gained}
\keyword{has}
\keyword{if}
\keyword{in}
\keyword{information}
\keyword{information).}
\keyword{input}
\keyword{internal}
\keyword{is}
\keyword{length}
\keyword{no}
\keyword{objects}
\keyword{over}
\keyword{same}
\keyword{scalar}
\keyword{simpler}
\keyword{so,}
\keyword{sub-daily}
\keyword{the}
\keyword{then}
\keyword{there}
\keyword{to}
\keyword{using}
\keyword{value}
\keyword{vector.}
\keyword{when}
\keyword{whether}
