% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_hla.R
\docType{data}
\name{UtilsHLA}
\alias{UtilsHLA}
\alias{utils_hla}
\title{Utility Functions for HLA Table ELT}
\description{
Collection of utility functions for HLA data
}
\keyword{HLA}
\keyword{Standardize}
\keyword{Table}
\keyword{an}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UtilsHLA-std_hla_tbl}{\code{UtilsHLA$std_hla_tbl()}}
\item \href{#method-UtilsHLA-std_hla_allele}{\code{UtilsHLA$std_hla_allele()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UtilsHLA-std_hla_tbl"></a>}}
\if{latex}{\out{\hypertarget{method-UtilsHLA-std_hla_tbl}{}}}
\subsection{Method \code{std_hla_tbl()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UtilsHLA$std_hla_tbl(tbl, na = c("", "NT", "Blank", "-", "Not Interpretable"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tbl}}{\verb{[data.frame]} HLA data}

\item{\code{na}}{\verb{[character]} Vector of values to interpret as \code{NA}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The standardized data
Standardize HLA Allele Representations
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UtilsHLA-std_hla_allele"></a>}}
\if{latex}{\out{\hypertarget{method-UtilsHLA-std_hla_allele}{}}}
\subsection{Method \code{std_hla_allele()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UtilsHLA$std_hla_allele(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{\verb{[character]} A vector of allele IDs}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Standardized alleles
}
}
}
