% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_mrd.R
\docType{data}
\name{UtilsMRD}
\alias{UtilsMRD}
\alias{utils_mrd}
\title{Utility Functions for MRD Table ELT}
\description{
Collection of utility functions for MRD data
}
\keyword{Chimerism}
\keyword{Data}
\keyword{Detect}
\keyword{Factor}
\keyword{Level}
\keyword{in}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UtilsMRD-str_detect_fct}{\code{UtilsMRD$str_detect_fct()}}
\item \href{#method-UtilsMRD-str_range_to_num}{\code{UtilsMRD$str_range_to_num()}}
\item \href{#method-UtilsMRD-str_excel_dt_to_num}{\code{UtilsMRD$str_excel_dt_to_num()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UtilsMRD-str_detect_fct"></a>}}
\if{latex}{\out{\hypertarget{method-UtilsMRD-str_detect_fct}{}}}
\subsection{Method \code{str_detect_fct()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UtilsMRD$str_detect_fct(x, lvl, lbl)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{\verb{[chr]} A character vector}

\item{\code{lvl}}{\verb{[chr(1)]} The level's numeric representation}

\item{\code{lbl}}{\verb{[chr(1)]} The level's label}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\verb{[lgl]} A logical indicating presence or absence of the indicated
factor level
Convert a \% Range to It's Midpoint
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UtilsMRD-str_range_to_num"></a>}}
\if{latex}{\out{\hypertarget{method-UtilsMRD-str_range_to_num}{}}}
\subsection{Method \code{str_range_to_num()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UtilsMRD$str_range_to_num(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{\verb{[character]} Character representation of a range of percentages}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\verb{[numeric]} The midpoint of each range of values
Convert Excel Date to Number

Converts Excel dates in MDY format to a number using an origin of
"1/0/1900". Elements that cannot be converted are returned as-is.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UtilsMRD-str_excel_dt_to_num"></a>}}
\if{latex}{\out{\hypertarget{method-UtilsMRD-str_excel_dt_to_num}{}}}
\subsection{Method \code{str_excel_dt_to_num()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UtilsMRD$str_excel_dt_to_num(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{\verb{[chr]} Character vector containing Excel dates}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\verb{[chr]} Numeric representation of Excel date, where available
}
}
}
